Removed codec comparison test: it didn't work and probably never will.
The central problem is that we cannot sync the frames in the input video with the output video, which makes PSNR/SSIM go crazy. The test only appeared to succeed earlier due to a bug in the test. We can consider this a failed experiment, but we did learn a lot from it :)

BUG=550

Review URL: https://webrtc-codereview.appspot.com/969006

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 849fb8ebd28986e7111190348f3f08183258b092