Reduce flakiness in network down test.
The encoder is in the process of encoding when the network goes down, so we need to wait until it has finished before we expect no more packets to be sent.

Also fixed a test which was testing the wrong thing.

R=mflodman@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/2258008

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 269dd4264f1edf62fff78c6bce57d9ffdf978a9b